WIP: replacing kdcount with hfof.
- need to investigate if this really is faster.

- a few pending changes; such that we can package hfof on bccp.

- test cases in nbodykit passes, but need to check if they are
  sufficiently powerful; perhaps either make them tighter or
  otherwise run it on a real dataset.

the single to double promotion is especially worrying in memory consumption.
there is probably a way to deal with that.

https://github.com/pec27/hfof/pulls/1

https://github.com/pec27/hfof/issues/2

https://github.com/pec27/hfof/issues/3

The true speed up will only be measured on real data sets. The slow parallel
merge algorithm here may be the dominating factor already; so the benefit
of the switch still needs to be investigated.